tcp sequence arithmetic: wrap int32 diffs/sums in (uint32) casts (UBSan UB)
Signed-overflow UB on LP64 when sequence fields (kept as int32 for
compatibility) overflow around the 2^31 boundary. Apply the
(uint32) cast idiom throughout tcpin.c/tcpsubr.c/tcpuser.c/tcpcmd.c/
tcpout.c and ftpcli.c diagnostics. geniss() casts msclock()<<12
through uint32 first to avoid signed-shift UB. Live-verified with
slip + vjslip/drop and netrom/flexnet rounds, UBSan-clean.

+++ src/ftpcli.c
@@ -584,7 +584,7 @@ static void ftpcds(struct tcb *tcb, enum tcp_state old, enum tcp_state new)
 			ftp->state = COMMAND_STATE;
 			if(Current != NULL && Current->cb.ftp == ftp){
 				printf("Put complete, %lu bytes sent\n",
-					(unsigned long)(tcb->snd.una - tcb->iss - 2));
+					(unsigned long)((uint32)tcb->snd.una - (uint32)tcb->iss - 2));
 			}
 		}
 		break;
@@ -598,7 +598,7 @@ static void ftpcds(struct tcb *tcb, enum tcp_state old, enum tcp_state new)
 			ftp->state = COMMAND_STATE;
 			if(Current != NULL && Current->cb.ftp == ftp){
 				printf("Get complete, %lu bytes received\n",
-					(unsigned long)(tcb->rcv.nxt - tcb->irs - 2));
+					(unsigned long)((uint32)tcb->rcv.nxt - (uint32)tcb->irs - 2));
 			}
 		}
 		break;
